BasicWeaponsAdded

+++ Assets/Scripts/ScriptableObjects/SOEnemies.cs
@@ -3,6 +3,6 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "SOEnemies", menuName = "Scriptable Objects/SOEnemies")]
 public class SOEnemies : ScriptableObject
 {
-    public int MaxHP;
+    public float MaxHP;
     public GameObject prefab;
 }
